(PUP-8598) Add configuration of loaders to agent application
The agent application uses the configurer to define the current
environment. It did so without configuring any loaders. This led to
that Binary file content would not work even if --rich_data was turned
on because it was not possible to find the "Binary" data type.

Now, loaders are configured when the current_environment is configured
by the configurer.