(PUP-8640) Add resolution of Deferred values in agent and apply
This adds resolution of all Deferred values present in a catalog before
apply of the catalog proceeds. The replacement is recursive such
that any Deferred values in Array, Hash, or Sensitive values are
replaced by a copy where a deferred value has been resolved.

The resolution transforms arguments to Deferred values before calling
the resolution.

In order to do this, the fact_handler was modified to enable getting
hold of the facts (to enable making them available to called deferred
functions).