(PUP-8676) Add a cache_param to InternalFunction - 4x function API
Before this, if a user wanted to cache information for one or a group of
functions this would require defining an adapter and adapting the
compiler (typically) or environment.

Now, a function that is an InternalFunction can declare a parameter as
`cache_param` and it will get an injected instance of an
`ObjectIdCacheAdapter` that responds to `retrieve(o)` where `o` is the
object the cache is associated with. A mutable hash is returned from
`retrieve`. The first time the hash will be empty.

The lifecycle of the cache is the same as that of the compiler.