(T.B.D) Fix mismatch sym/str in StaticDirectory and faulty modulepath
The StaticDirectory env did not work when trying to load an
environment config since there was a mismatch between expectation
to compare env name as symbol and name being a string.
This is fixed in this commit by making the name of the env be a
symbol. (This may not be the right way to fix this - other
alternatives created a symbol name when comparing).

This also changes setting the default modulepath to an empty
array instead of an empty string. Apparently this functionality
has never been tested since the rest of the system expects it
to be an array. This changes the default module path to an empty
array.