(w.i.p) Change fix to use args for files instead of single -f
Before this a file with reported issues was given with an
option -f followed by filename. This was not great when having multiple
files.

Now this is changed so that files are given as arguments without any
options.

This commit also makes it legal to specify a single issue using
`--issue` in combination with issue files; the single issue is
simply given to the set of reported issues.