(w.i.p) Add combination of reported issues
If there are several entries reporting that an issue is present
on a node all such entries are now combined to the same output.

This is useful when a scanner scans nodes individually and each
result is transformed to an individual puppet fix report. Now, when
multiple files are given to puppet fix it will just perform one
call (run_task/run_plan etc.) for all nodes across all reports.